feat(core): every write copies its value into a tree wlog owns (CORE1)
value.go holds the copy. The tree holds only nil, bool, string, int64, uint64,
float64, json.Number, map[string]any, and []any, and it follows the json field
rules of encoding/json: tags, omitempty, "-", the string option, and embedded
structs. The conversions follow the spec: an error becomes its message, a
duration becomes milliseconds, a time becomes RFC 3339 UTC, raw JSON is decoded,
other bytes become their length, and a NaN or an infinity becomes its name. A
value that cannot be copied becomes "[unencodable: type]", and nesting past
depth 16 becomes "[truncated: depth]".

The copy runs before the event lock is taken, so a MarshalJSON method that logs
through wlog on the same event finishes instead of deadlocking. A typed nil
error and a user type that panics in Error meet the same guard.

The tests name the finding they close: CORE2, CORE3, CORE4, CORE5, CORE6, and
CORE26. The tree stores int64 rather than int, so the tests that compared a
stored number with an int now compare with int64.

+++ llm/record_test.go
@@ -53,9 +53,9 @@ func TestLLM_Set(t *testing.T) {
 	group := llmGroup(t, rec)
 	want := map[string]any{
 		"provider": "openai", "model": "gpt-x", "operation": "chat",
-		"input_tokens": 10, "output_tokens": 5, "cached_input_tokens": 3,
-		"reasoning_tokens": 2, "total_tokens": 15,
-		"tool_call_count": 1, "tool_call_failures": 1,
+		"input_tokens": int64(10), "output_tokens": int64(5), "cached_input_tokens": int64(3),
+		"reasoning_tokens": int64(2), "total_tokens": int64(15),
+		"tool_call_count": int64(1), "tool_call_failures": int64(1),
 		"time_to_first_chunk_ms": int64(100), "duration_ms": int64(2000),
 		"streamed": true, "finish_reason": "stop",
 		"cost_micros": int64(300),
@@ -89,14 +89,14 @@ func TestLLM_Add_FoldsTotals(t *testing.T) {
 	end()
 
 	group := llmGroup(t, rec)
-	if group["input_tokens"] != 13 || group["output_tokens"] != 6 || group["total_tokens"] != 19 {
+	if group["input_tokens"] != int64(13) || group["output_tokens"] != int64(6) || group["total_tokens"] != int64(19) {
 		t.Errorf("folded tokens = %v/%v/%v, want 13/6/19", group["input_tokens"], group["output_tokens"], group["total_tokens"])
 	}
 	calls, _ := group["calls"].([]any)
 	if len(calls) != 2 {
 		t.Fatalf("calls = %v, want 2", group["calls"])
 	}
-	if group["tool_call_count"] != 2 || group["tool_call_failures"] != 1 {
+	if group["tool_call_count"] != int64(2) || group["tool_call_failures"] != int64(1) {
 		t.Errorf("tool call totals = %v/%v, want 2/1", group["tool_call_count"], group["tool_call_failures"])
 	}
 }
